move rss stuff in a separate subpackage
