+++ my/rss/feedbin.py
@@ -1,13 +1,13 @@
 """
-Feedly RSS reader
+Feedbin RSS reader
 """
 
-from my.config import feedly as config
+from my.config import feedbin as config
 
 from pathlib import Path
 from typing import Sequence
 
-from .core.common import listify, get_files, isoparse
+from ..core.common import listify, get_files, isoparse
 from ._rss import Subscription
 
 
@@ -24,25 +24,22 @@ from datetime import datetime
 def parse_file(f: Path):
     raw = json.loads(f.read_text())
     for r in raw:
-        # err, some even don't have website..
-        rid = r['id']
-        website = r.get('website', rid) # meh
         yield Subscription(
-            created_at=None,
+            created_at=isoparse(r['created_at']),
             title=r['title'],
-            url=website,
-            id=rid,
+            url=r['site_url'],
+            id=r['id'],
         )
 
 
 def get_states() -> Dict[datetime, List[Subscription]]:
-    import pytz
+    # meh
+    from dateutil.parser import isoparse # type: ignore
     res = {}
     for f in inputs():
+        # TODO ugh. depends on my naming. not sure if useful?
         dts = f.stem.split('_')[-1]
-        dt = datetime.strptime(dts, '%Y%m%d%H%M%S')
-        dt = pytz.utc.localize(dt)
+        dt = isoparse(dts)
         subs = parse_file(f)
         res[dt] = subs
-        # TODO get rid of these dts...
     return res
