+++ my/rss/feedly.py
@@ -1,13 +1,13 @@
 """
-Feedbin RSS reader
+Feedly RSS reader
 """
 
-from my.config import feedbin as config
+from my.config import feedly as config
 
 from pathlib import Path
 from typing import Sequence
 
-from .core.common import listify, get_files, isoparse
+from ..core.common import listify, get_files, isoparse
 from ._rss import Subscription
 
 
@@ -24,22 +24,25 @@ from datetime import datetime
 def parse_file(f: Path):
     raw = json.loads(f.read_text())
     for r in raw:
+        # err, some even don't have website..
+        rid = r['id']
+        website = r.get('website', rid) # meh
         yield Subscription(
-            created_at=isoparse(r['created_at']),
+            created_at=None,
             title=r['title'],
-            url=r['site_url'],
-            id=r['id'],
+            url=website,
+            id=rid,
         )
 
 
 def get_states() -> Dict[datetime, List[Subscription]]:
-    # meh
-    from dateutil.parser import isoparse # type: ignore
+    import pytz
     res = {}
     for f in inputs():
-        # TODO ugh. depends on my naming. not sure if useful?
         dts = f.stem.split('_')[-1]
-        dt = isoparse(dts)
+        dt = datetime.strptime(dts, '%Y%m%d%H%M%S')
+        dt = pytz.utc.localize(dt)
         subs = parse_file(f)
         res[dt] = subs
+        # TODO get rid of these dts...
     return res
